pyproject.toml
[tool.poetry]
name = "task-manager"
version = "0.1.0"
description = ""
authors = ["RocketMenace <grizzlygit777@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.116.1"}
sqlalchemy = "^2.0.42"
asyncpg = "^0.30.0"
alembic = "^1.16.4"
pydantic-settings = "^2.10.1"
dishka = "^1.6.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.12.7"
pre-commit = "^4.2.0"



[tool.poetry.group.test.dependencies]
pytest = "^8.4.1"
httpx = "^0.28.1"
pytest-asyncio = "^1.1.0"
pytest-cov = "^6.2.1"
faker = "^37.5.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
